requirements.txt
numpy>=1.26
scipy>=1.12
matplotlib>=3.8
Pillow>=10.0
tqdm>=4.66
